Refactor profile helpers and add gaussian_profile to carRecharge.py
Moved gaussian_profile and related profile helper functions from app.py to carRecharge.py and a new appHelper.py module to improve code organization and reusability. Updated imports in app.py to use the new helper module. Enhanced .vscode/settings.json for better Python analysis and auto-imports.
